fix(ci): Comply to PEP-668
On python dependencies install, there was a 'error: externally-managed-environment'
message.
A virtual environment was set to comply with the PEP-668 specification.
As the python inside the virtual environment could not find the `asciidoc`
module installed by apt, this commit uses pip to set it up.

Merge branch 'fix-pep-0668-error' into dev-short-system-tests