chunkserver, common: Add the FolderChunks class
The struct folder was using 'Chunk *testhead,**testtail and
uint32_t chunkcount' in combination with 'Chunk *testprev, *testnext'
to keep track on the chunks and to distribute the chunks to be tested.

This commit replaces the previous behavior with the new FolderChunks
class. The 'struct folder' now contains a instance of this class which
allows to:

- select a single chunk at random in constant time,
- shuffle all chunks into a random order,
- constant-time operations of insertion, deletion, mark as tested,
  and get-chunk-to-test.

This collection uses the 'indexInFolder' field in the Chunk structure
to be able to achieve constant-time operations.

The unit tests were also added.