chunkserver: Refactor struct cntcond and its use
'struct cntcond' was using C style linked list and raw pointers. The
structure name was also confusing.

This commit renames it to be 'struct CondVarWithWaitCount' and improve
the member variable names.

The member variable in the Chunk class and the list to reuse the unused
condition variables 'gFreeCondVars' are now using smart pointers.