chunkserver: Refactor Folder::todel variable
The 'todel' variable was holding 0, 1 or 2 with different meanings,
then its usage was confusing and hard to read, representing sometimes
a read-only state and sometimes also marked for removal from the
config file (hdd.cfg).

This commit splits this variable into isReadOnly and
isMarkedForRemoval, and refactors the usage on hddspacemgr.

The 'todel' variable belonging to the Chunk class is not needed
anymore, it is inherited from its owner folder.

Co-authored-by: Ruben <ruben.crash@gmail.com>
Co-authored-by: Baldor <baldor@leil.io>
Co-authored-by: Antuan <aaim@leil.io>