protocol, cgi, chunkserver: Remove unused packets
Protocol packets CLTOCS_HDD_LIST_V1 and CSTOCL_HDD_LIST_V1 were kept
for backward compatibility with LizardFS versions lesser or equal than
v1.6.8. The only module using them was the monitoring.

As we do not plan to support such old versions, it is safe to remove
the packet definitions and the death code that depends on it.