chunkserver: Refactor Folder class usage
Previously, all the fields needed to be initialized after the creation
of a folder instance, even if it could be filled with default values.
Also the 'path' was stored in a char *, hence we needed to free it
explicitly.

This commit adds the constructor with the mandatory parameters and
assigns correct default values. The field 'path' was also changed to
be a std::string and the hdd_parseline function was changed
accordingly.

Other minor changes about fields organization, comments and style.

Co-authored-by: Ruben <crash@leil.io>