chunkserver: Move disk info logic to Folder
The logic of function hdd_diskinfo_v2_data was completely on the
hddspacemgr, even if the information was composed by the Folder's
properties.

This commit moves the responsability of constructing the DiskInfo
objects to the struct Folder. hddspacemgr only needs to iterate now
trough the list of Folders and call the function.

As the new folder.cc file was created, we can move the definition of
the Folder's functions to this unit.