nfs-ganesha: Refactor code to remove liz prefix
The struct liz_fileinfo is declared outside this code at the
lizardfs_c_api library. So a typedef alias is declared to remove
the branding dependency in order to make it more generic.

A new camelCase namming convention is applied to match other modules
coding style. The style is applied on functions and types.

Doxygen code style comments are added.

Co-authored-by: Crash <crash@leil.io>