docs: Add Doxygen documentation for nfs-ganesha
The Ganesha plugin is still not completely documented. There are some
methods with a signature not properly described for being easily
read and understood by other members of the team.

The following commit adds Doxygen documentation for methods related
with the initialization of the FSAL and the creation and handling of
exports.