refactor: Refactor nfs-ganesha
This commit renames the method for creating the FSAL context used by
Ganesha when performing the filesystem operations. The previous name
didn't use the name convention used by the refactoring and the new
name is more appropriate.

This commits also removes some unnecessary whitespaces in the
signature of some methods.