fix(ci): Remove error message from build log
The docker image for cache use is expected to be missing from the
registry when it has not been created yet. Thus, it should not raise
an error. Previously it was not stopping the pipeline, but the log was
showing an error message in Blue Ocean, which does not deserve
attention, while something else after is the actual issue. This commit
is addressing that confusing situation.