master: Add dumping mount configurations
Mounts are configured a little bit differently than most of the
configurations. First is that it has a special parser different from the
common one. Second, in addition to parsing key-values, it also parses
FUSE command line options and mount points.

This commit gathers all the non-comment lines into a string. This string
will be used when the mount is registering to the master. Master will
save this info into the session struct, which will be kept until master
is stopped or the mounts disconnect (after 15 seconds at time of commit).

Also fixes some FUSE problems, particulary that the old FUSE headers
were still being included and causing problems in CI. There is still an
issue with the old FUSE headers still being included somewhere, so a
small workaround for this has been added in common/platform.h.