mount: Return ESTALE if ENOENT in sfs_getattr
In a very specific case, if a --bind mount is made into a SaunaFS mount
sub-directory, and it is deleted, sfsmaster will report to the bind
mountpoint that it doesn't exist. In other networking filesystems like
NFS, a stale file handle is reported instead. This could cause failures
in things like kubelet in failing to remove bind mount points.

This commit will return ESTALE in sfs_getattr if the file or directory
is not found. We can probably safely assume that the inode FUSE/Linux is
using doesn't exist anymore and someone is using a stale file handle.