feat(master): Allow change metadata save period
By default metadata dump occurs every hour at full hours. This
commit creates a configuration that allows to change the default
interval for dumps. Also, by setting it to 0 it is possible to disable
the dump (not recommended in production). There was a secret config
variable for disabling this, that has been superseded by this commit's
new config.

Co-authored-by: Antuan <antuan@leil.io>