master: Allow dumping default values
Previous iteration of the configuration dumping command did not allow
dumping default values of services, since default values are set by the
caller to the cfg_get* functions, thus making it non-trivial.

This commit adds a new header that contains most of the default
configuration options for master, shadow, chunkserver and metalogger and
creates a map for each service supported. It adds a --defaults option to
the dump-config command, which allows printing these default options.

This isn't perfect, some options are still missing and there's a bug
where it shows shadow's configuration options even if it's not running.
However, this is a big step in centralizing the configuration
management. In a potential future commit, instead of writing to a manual
page the options, it could be generated from the header file.