test: Fix Ganesha multiple export test
The previous version of the test validated that multiple NFS exports
can be shared simultaneously with the same Ganesha config file. This
test also validated the use of different versions of nfs protocol like
v4 and v4.1. However, nfs client started failing to connect to NFS
shares using versions 4.1 and 4.2. The client failed with the message
`mount.nfs: Protocol not supported`. The test started failing after
upgrading nfs-common to v2.6.2, but also failed in CI with nfs-common
v2.6.1.

This commit removes the option `-o v4.1` that specify the nfs client
to connect using nfs version 4.1. After this change, nfs client uses
nfs version 4 to connect and the test passes successfully.