admin: Add --defaults options
Previous iteration of the configuration dumping command did not allow
dumping default values of services, since default values are set by the
caller to the cfg_get* functions, thus there being no single truth to
these option values or even names.

This commit adds maps of the currently known configuration options, and
the dump-config command uses these to figure out what are the defaults
for each service.

This is done with best-effort, because this list may become outdated as
time goes on and new options are added. Until we have a single source of
truth for these options, this will have to do.