chunkserver: Make symbols visible in plugins
Some symbols needed for the defragmentation were not visible from the
plugins. This commit moves them to the chunkserver-common library.

The flag CMAKE_ENABLE_EXPORTS was added to share the same symbol
references between the chunkserver and the plugins. This is needed
because the plugins are shared libraries. If this variable is not set
or OFF, then the same global, probably static, variable will be
different on each executable (chunkserver and plugin).

The call to the defragmentation was removed from the tester thread, it
is now responsibility of the concrete plugin.

Co-authored-by: Crash <crash@leil.io>
Co-authored-by: guillex <guillex@leil.io>