all: Add option to generate compile_commands.json
Clangd and vscode rely on this file to understand the build context of
each file in the project. This commit adds the option to easily
generate it. The option is enabled by default.

To properly work, the file should be in the root of the repository.
At this point, a manual soft link should be generated, pointing to the
actual file (cmake_build_directory/compile_commands.json).