fix: Fix Windows compilation of common library
Latest changes in the dev branch brought the memory_mapped_file.cc and
the scoped_timer.cc files to the common library. These two files do not
compile under Windows due to the presence of the includes sys/mman.h and
sys/syslog.h. Solution is to add it to the list of excluded files in the
CMakeLists of the common library.