fuse, mount: Add DANGEROUS_sfsignoreflush option
The DANGEROUS_sfsignoreflush option targets answering inmediately to
the flush or fsync syscall without properly checking its responses. The
core of implementation of both syscall is wriite_data_flush function,
this is why is the mocked one. This fools single-file copiers into
believing the files are already copied, and the system can handle those
copies in a multithreaded fashion.