chunkserver: Fix global variables visibility
The static variable gPerformSync was not shared correctly between the
chunkserver and the chunkserver-common at runtime. Causing different
values from each side.

This commit makes it inline, which also fixes the
ShortSystemTest.test_fsync_eio_in_chunkserver.

The variable gScansInProgress was also changed to be inline to avoid
the same issue.