mount: Use inode_multiset_ to remove duplicates
Consider the case some entries has already been inserted with some
partial information, and those entries should present in the
inode_multiset_ member of the DirEntryCache, but not necessarily in any
of the other sets (index and lookup sets). So, if an entry with same
pair <inode, context> is once more inserted to the cache then the
inode_multiset_ would have two entries with the same <inode, context>
pair which may lead to not answering the latest inserted data to some
requests, for instance the getattr call on DirEntryCache::lookup.