mount (win): Fix .saunafs_tweaks behavior
Attempting to update the value of a variable from the `.saunafs_tweaks`
file under some Windows machines has been unsuccessful due to a read
operation being executed after the corresponding write operation.
This resulted in the changes from the write operation not being
reflected because the `getAllValues` function, which resets the
`.saunafs_tweaks` values to their defaults and did not check if a write
operation had been executed.

This issue was resolved by adding a check in the read operation
function related to `.saunafs_tweaks`. If the file had been written to,
a procedure is initiated that constructs the expected string format and
applies the correct update operation on the `.saunafs_tweaks` data,
according to the format of the input from PowerShell commands.

Co-authored-by: Dave <dave@leil.io>