fix: Consider 'no more entries' marker case
The 'no more entries' marker case is the situation when the readdir
operation is finished caching entries to the DirEntryCache. This line
```gDirEntryCache.insert(ctx, ino, 0, marker_index, marker_index, "", Attributes{{}}, data_acquire_time);```
is called and gives the idea of the content of such an entry. It means
that there is an entry which was not being inserted to the lookup_set_,
but was indeed being inserted to the index_set_ so it was breaking out
consistency checks. Solution was to include these 'no more entries'
markers in the lookup set.