fix: Reply newest data in inode-wise lookup
Some tests that check metadata recovery were failing by returning old
```mtime``` values. The root cause of these old values to be replied was
the presence of old entries in the inode_multiset_ of the DirEntryCache
and the returning of the data from these old entries. Solution was to
find the newest data from inode_multiset_ and reply it.