mount: Invalidate also the inode in link function
Previously, only the parent was invalidated from the cache, it was
working because the items were inserted in less places, so this event
was not happening. As we are now caching in lookup and getattr it is
expected to use more invalidations.

Makes the test GaneshaTests.test_nfs_ganesha_conectathon_nfs_suite to
pass the link stage, but it is still failing in the unlink test.

Co-authored-by: Crash <crash@leil.io>