all: Fix build under ENABLE_TRACES (#42)
Some of the calls to the traces macros were not updated in previous
commits. This change fixes the building by updating these calls with
correct parameters and extends the TracePrinter to optionally output
to our usual syslog.

To use the traces, just enable the option ENABLE_TRACES in the cmake
call. By default it prints to stdout, so, to print to the syslog, the
option ENABLE_SYSLOG_FOR_TRACES must be enabled too (useful when
running the tests).

To disable the traces for specific modules, add the following line in
the module cmake file:

remove_definitions(-DENABLE_SYSLOG_FOR_TRACES)