feat(tests): Add core dump watch and analysis
This commit aims to improve the behavior and improve the output of test
results. When a core dump occurs, it terminates the test immediately
(regardless whether it was SaunaFS or not that caused it, the
environment is unreliable for accurate test results). It then uses
GDB to analyze the core dump(s).

There's a few issues with this: First it requires modifying
`/proc/sys/kernel/core_pattern`, which will affect the whole system.
While I've tried to ensure that the original pattern is restored after
failures and core dumps, I'm not completely confident it will. Second is
the fact that GDB might not be needed to print the backtrace:
`LD_PRELOAD=/lib/libSegFault.so` may be a better alternative, but I
didn't have time to test it.

Thus this feature is hidden behind a feature flag as experimental until
these issues are solved.

Co-authored-by: aNeutrino <piotr.modrzyk@gmail.com>