fix(tests): Upgrade tests (#58)
This is a series of 2 commits squashed:

fix(tests): Upgrade tests on freshly installed machines

This commit fixes an issue from earlier regarding updates to
setup_machine.sh. Now that the SaunaFS key is acquired from the
setup_machine.sh script and it adds the repository, there is no need to
add a source to the script when doing upgrade tests. Doing so would
conflict with the already setup repository

fix(tests): Long upgrade tests

* Fix no old version checking by using latest client instead.
* Fix recursion bug by not using sfs* tools, instead using saunafs