fix(tests): Fix various issues in the testing framework (#55)
* Remove unnecessary ticks
This messes up some editor syntax highlighting
* Fix binary output being outputted
The variable used in the conditional is different from the actual
filename used. This conditional would always succeed since awk would
return 0 if the conditional in awk would not match, which it didn't
since stdout was empty (only stderr was populated with the error
message, but it is disregarded).
* Fix coredumps not being generated
At least 3 years ago, `sudo` had a change of behavior regarding
coredumps. This fixes that by setting the default for cores as the
system.