mount: Reduce readings during EC write (#54)
Tests performed show that during EC writes the client connects and asks
to the chunkservers for the missing data parts of the current stripe,
which are the data parts the client are not writing. This is required if
the file is bigger than the last byte it is writing, but it should not
ask for data on offsets beyond the file size. It is specially important
for the writings of small files, and removing those extra requests
should reduce overhead.

Current solution manages to calculate chunkSizeInBlocks_ for each
ChunkWriter at the moment of initialization from the field maxfleng of
the current inode's inodedata. Thus, the maxfleng field should be
updated the best way possible. The chunkSizeInBlocks_ is then used to
check if client ws about to was to chunkservers for not existing data.

Co-authored-by: Dave <dave@leil.io>
Co-authored-by: guillex <guillex@leil.io>
Co-authored-by: Roly <roly@leil.io>