chore: Add .editorconfig (#50)
.editorconfig allows signalling to code editors how files should be
formatted when opening and when writing. For example, since most of us
uses 4-width tabs, the default should be to use those (especially
because tab sizes can vary, from 2 to 8 to any size). This can be
modified per filetype. Another thing is to ensure that files should end
with LF, not CRLF.