admin: Deprecate saunafs-probe command (#26)
* admin: Deprecate saunafs-probe command (step 1)

The command was introduced in version 2.5.0 (2014) and later renamed
from 'probe' to 'admin' (2015). A soft link is kept for backward
compatibility. Even if both work, there were some inconsistencies,
like:

- The package name was admin, but all the main classes still had probe
  in the name.
- Some of the tests had admin in the filename while other had probe.
- They look like different commands but the other is just a soft link.

This commit starts the deprecation of the saunafs-probe command by
removing most of the references from the tools, tests and C++ code. The
soft link, packages and documentation was kept for backward
compatibility and will be removed for version 5.0.0.

* tidy(docs): Add note about saunafs-probe

---------

Co-authored-by: Urmas Rist <urmas@leil.io>