chore: Update CONTRIBUTING.md and README.md
Previously we said we won't accept outside contributions and while the
policy hasn't changed per-se, changing it is not a priority. To not
discourage people from submitting code, the files are updated ask people
to contact us before contributing any code to us, so we have time to
respond and work out the necesarry details needed to accept the code.

Fix an empty link in README.md as well.

Closes #8