mount: Add sfsignoreflush option (#34)
The sfsignoreflush option targets answering inmediately to
the flush or fsync syscall without waiting for the responses. The
core of implementation of both syscall is the write_data_flush
function, this is why is the mocked one.

This fools single-file copiers into believing the files are already
copied, and the system can handle those copies in a multithreaded
fashion.

Co-authored-by: Dave <dave@leil.io>