test: Fix connectathon test in Ganesha v6
The previous version of the test cloned cthon repository on top of a
NFS mount exported through Ganesha before building the project and
running cthon suite. After upgrading to Ganesha v6, git clone command
started failing with permission denied error and therefore the test
fails.

This commit applies a temporary fix for cloning cthon repository on
top of a SaunaFS mount. With this change, even if cthon is cloned in
a different path, cthon suite keeps running on top of a Ganesha mount
and the test passes successfully.

This patch will be restored after finding the solution for this issue.