chunkserver: Increase wait time for locked chunk
The constant kSecondsToWaitForLockedChunk_ was introduced to avoid
avoid and point out possible deadlocks, by waiting the for specified
time to release the condition variable waiting on a Chunk.

The previous value (2 seconds) was too small for testing scenarios
with very slow disks or a high number of operations on the same disk.

The new value (20 seconds) is small enough to not stop the IO operation
(due to the mount retries) and big enough to detect unusual/unexpected
behavior.