mount: Fix test_sparse_file (#77)
Recent runs of this test under Windows failed apparently due to
different reasons. Sometimes during a setattr syscall and sometimes
during a write syscall. Debug revealed weird values of the maxfleng
field when it was supposed to be zero.

The cause of this issue was that the inodedata (```id```) could be
released in the ```write_data_lcnt_decrease``` function and the later
update of the maxfleng field was pointing to already freed memory.
Solution consisted in checking whether the inodedata was released and
only perform the update of its maxfleng if it was not deleted.

Co-authored-by: Dave <dave@leil.io>