feat(metrics): Add initial prometheus support
This commit introduces some initial prometheus support. This is both a
compile and runtime feature. First you must have the prometheus-cpp
library installed for it to compile support for this feature. Second,
you must set ENABLE_PROMETHEUS to 1 in sfsmaster.cfg.

Currently prometheus is only implemented for master, and is simply
collecting packet data sent and received from client. In the future,
more metric data can be added, and hopefully most of CGI metrics can be
replaced with this.

This also introduces some initial support for multi-threading in master
(for prometheus) and adds documentation regarding this new feature.