mount: Apply fixes to DirEntryCache
Recent tests added to the DirEntryCache batch of unittests revealed some
bugs in its implementation. One of them consisted in erasing more
"parent" entries on both lockAndInvalidateParent implementations. The
other issue was an early stop of the first (and primary) while of the
lockAndInvalidateParent implementation which considers contexts.

Debugging sessions on failing tests showed issues with files with
hardlinks, specially caused by the imposibility of knowing the inode of
the object to remove in the unlink syscall (without extra communication
with master) which could be used to call lockAndInvalidateInode. Some
ideas were tried, but the other ones seems to be as bad as any. This one
was chosen due to it only affects the files with hardlinks, only with
negative effect it should be asked some extra times its metadata to
master in sequential queries for data. This seems logical due to the
previously presented issue.