tests: Clean up space used by the tests
Previously, after running the tests, used space was not cleaned because
the path for disks and loop devices was not correctly setup.
Therefore, the used space by the tests needed to be cleaned manually to
avoid the disks become full.

This commit updates the path for disks and loop devices including data
and meta folders. Consequently, the space used by the tests is now
cleaned up correctly.