fix(setup): Remove apt repository setup
The setup machine script was failing on apt update caused by access
issues to the apt repository. As this is required to pass for further
creating the packages that need to be in the repository, it created a
chicken-egg problem. This commit removes the setup of the repository
that was introduced recently on that script.