feat(chunkserver): Add DiskManager interface
Previously, the disk assignment was done in a function inside the
hddspacemgr.cc file, which works fine for basic placement of chunks in
the disks, but is not enough if an advanced or another specific
distribution is required.

This commit introduces the IDiskManager interface and the
DefaultDiskManager concrete implementation, to provide the
extension points for a future more complex disks management system.

So far, the only virtual function is getDiskForNewChunk, but others
will appear soon.

The DefaultDiskManager provides the same distribution as before.

Co-authored-by: Crash <crash@leil.io>