feat(chunkserver): Introduce DiskEnergyManager
Previously, the disk assignment was done in a function inside the
hddspacemgr.cc file, which works fine for basic placement of chunks in
the disks, but is not enough if an advanced power management system
is required.

This commit introduces the IDiskEnergyManager interface and the
DefaultDiskEnergyManager concrete implementation, to provide the
extension points for a future more complex energy management system.

So far, the only virtual function is getDiskForNewChunk, but others
will appear soon.

The DefaultDiskEnergyManager provides the same distribution as before.