fix(tests): Race condition in metadata_dump_period_seconds (#110)
In rare cases, if master starts up at the same second as a metadata dump
should occur, the metadata dump may not happen. However, since the last
seconds digit may still be 0 or 5, it will skip waiting for the metadata
dump and continue on, resulting in failure.
These changes fixes this issue by first syncing up with the first
metadata dump and then waiting a second before the actual tests begin.
It also adds echo statements to better track the time.