feat(common): Improve the hddSizeParse function (#129)
The previous version was completely implemented from the scratch. The
proposed version uses modern C++ and std::stod to increase readability
and safety. The commit makes it also more flexible with different
suffixes for the size and improves the error handling.

The function initially was in the hddspacemgr.cc in the chunkserver,
but it is very general and will be used from the plugins soon, so it
makes sense to move it to the common library.

An unit test file was also added to verify different combinations.