feat(uraft)!: Add warning and no-ip option
Previously, if you didn't run the program as root, it would ask for a
password. The previous workaround for this was to add sudo permissions
for the user to run the `ip` program. This commit changes this
behaviour, requiring root for normal uraft behaviour, with exceptions of
elector mode and a new option: --no-ip

--no-ip is useful for local testing and debugging. If you are testing
locally with multiple ports, you may have no desire to add floating IP
addresses (since all share the same machine). This is of course, no use
in production.

BREAKING CHANGE: Explicitly requires root to run the uraft program under
normal circumstances.