feat(mount): Random case option parsing
For allowing the client users to input option with random case formats,
this change allowing that on the client argument processing.

It also enables the parsing of the following formats of options
written in the cfg file (if provided, obviusly): <-D value>-like
format, this is a single character keyword like -P, -D, -B or -S and
<-o opt1=val1,opt2=val2,...> format.

Plus, it enables the random casing in the names of the options and
forbids mount if cfg is provided and there are some unexpected (like
--help, -V, -u options) or wrong options.

The proposed changes enforce cfg files to be well-formed, thus causing
the mount to fail in such cases. Some tests were fixed due to having
such errors.

Two integration tests were also added:
- a test that check the expected random casing parsing in cfg files.
- another test to ensure the behavior of random case options parsing
from command line work properly.

Co-authored-by: Dave <dave@leil.io>