feat(metrics): Add rest of master chartdata metrics
This commit finishes adding metric data from master chartsdata. Since
all the chartsdata was counters in Prometheus terminology, there was no
need to add other types, however a a small refactor was needed to
accomodate master metrics and in the future other service metrics.
Notably removed was CPU usage and memory, since there are better ways to
gather this data, including with a prometheus server itself (which will
scrape this data).

This commit deprecates master chartsdata, and is the first step in
deprecating the CGI monitoring