feat: Load legacy metadata file
The master server now supports loading legacy `metadata.mfs` file. The
new logic priority is to search for new metadata format and loaded if it
is found. If not found it search for legacy format, load it and save it
in new format. If both are present, it load new metadata `metadata.sfs`.

- fix: Apply code review suggestions