fix(tests): Make Short and Sanity work with mfschunkserver
version: 3.13.0-devel-32e017505b4d0e994bc9dc47a641149b6b26bf1e

32e017505b4d0e994bc9dc47a641149b6b26bf1e

Those tests are needed to confirm that legacy chunkserver
is working well with the rest of cluster using SaunaFS binaries

We need it in order to dertermine which version of SaunaFS is still
working with legacy Chunkserver.

Mixing legacy chunkserver with SaunaFS cluster is one of the steps
in the migration process from legacy systems.

Co-authored-by: guillex <guillex@leil.io>