fix(mount): Fix IIS permissions mess (#154)
fix(mount): Fix IIS permissions mess

This commit aims to provide the expected behavior of our system while
dealing with the IIS tool. This tool apparently uses a very specific
user (17) to access the underlaying files. The user 17 was not expected
to perform these kinds of operations, thus causing the permission denied
errors. Solution was to make a breach at some checks to allow it using
the UIDs provided by the ```sfsallowedusers``` option.