test: Add sfstests
sfstests is a Go program designed to run tests in parallel and locally.
It uses Docker containers to run a image with a SaunaFS testing
environment, which allows isolated testing environments. It
automatically creates and destroys these containers.

A Github action is added for running Sanity and Short tests to show how
this works in action and as an alternative to help community
contributors run tests.

In the future, this could potentially saunafs-tests binary, as it allows
far more customization as needed. Currently it uses it to run the tests,
but this could be skipped and the bash scripts invoked directly from
sfstests.