fix(mount): Fix client race conditions
This commit targets getting rid of race conditions in the client side
with the help of helgring (valgrind suite). Though it does not remove
all of the issues pointed out by helgrind, it does fixes many of them.

List of changes:
- rename operations manager mutex to gReadaheadRequestsContainerMutex
and restrict its use to the cases when the container is used.
- add mutex per ReadRecord to protect its readaheadRequests,
readahead_adviser and cache.
- add mutex per read cache's Entry to protect its members, specially
the buffer and the timer.
- fix issue that may cause crash in the client related with the proper release of the entries.
- remove ReadRecord's locator and replace it by creating locators per
ChunkReader.
- protect terminate variable from masterproxy.
- protect inodedata's maxfleng access in writedata.