fix(mount): Fix many parallel reads issue
Root cause of the issue was that upcoming readhead requests were
turning done in the middle of the two calls to ReadCache::query
function in request function. Those requests were being processed
by the read workers and the chenges were protected by the entry's
mutex, meanwhile the request is under the protection of the fileinfo's
and inode (ReadRecord's) locks.

Solution is to check whether the expected undone entry (from the second
query call) is done or not, and just before returning check if there is
some undone entry to return whether or not wait for something.