refactor(chunkserver): Use list + smart pointers
ChunkserverEntry was using a C-like linked list like this:
- PacketStruct *outputHead = nullptr;
- PacketStruct **outputTail = &outputHead;
This change replaces it with a std::list:
- std::list<std::unique_ptr<PacketStruct>> outputPackets;

The commit also uses smart pointers for the list elements and for
the members readPacket and writePacket.

Some of the related functions were moved (and renamed) to the
ChunkserverEntry struct, where they look more natural.