feat(master): Add option to toggle data parts priority (#208)
Originally, the master server was placing the data parts in the
chunkservers with higher available space factor. Then, the parity parts
were placed in the remaining chunkservers.

If the cluster is not balanced (some chunkservers have more space than
others), then almost all parities land in the same chunkservers. This
causes that reading is always issued to the chunkservers with data
parts, so the chunkservers with parity parts are barely used when
reading.

This commit introduces the configuration variable PRIORITIZE_DATA_PARTS
to control this behavior. The default value is 1 (active), which is the
same as it was before.