feat(uraft): Add mechanism to recover floating IP
Previously, the loss of the uRaft floating IP caused SaunaFS services
to stop. Consequently, the cluster was unable to recover and it was
required a manual restart of the uraft service to bring back the
floating IP.

This commit introduces a recovery mechanism that allows SaunaFS to
regain the floating IP after unexpected failures.

The key changes are:
- Added `saunafs_is_floating_ip_present()` in `saunafs-uraft-helper` to
  check the current state of the floating IP.
- Added `isFloatingIpAlive()` in uRaft to check the floating IP status
  by invoking a helper script function.
- Introduced logic to demote the current leader if the floating IP is
  lost and ensure metadata synchronization between shadow servers and
  the newly elected leader. It avoids having split brain scenarios.

Co-authored-by: Baldor <baldor@leil.io>