feat(master): Improve openedfiles per session
These changes target improving the performance of sessions which have
already opened many files. Previous implementation of the openedfiles
member was a linked list sorted by inode which caused that searching,
inserting, counting how many files are open and most operations were
spending time proportional to the already opened files, i.e O(n) if n
is the number of opened files.

The solution proposed changes the type of this member to std::set and
makes complexity of all operations O(logn).

This change also improves readability of this part of the code.