feat: Improve acquired files in client
These changes target lowering the CPU load of clients which have
already opened many files. Previous implementation of the afhead global
variable was a linked list sorted by inode which caused that searching,
inserting, counting how many files are open and most operations were
spending time proportional to the already opened files, i.e O(n) if n
is the number of opened files.

The solution proposed changes the type of this variable to std::map and
makes complexity of all operations O(logn).

This change also improves readability of this part of the code.