feat: Improve idhash in client
These changes target lowering the CPU load of clients which are writing
on many files. Previous implementation of the idhash variable was a
linked list per each different hash which caused that searching,
inserting, counting how many files are open and most operations were
spending time approximately proportional to the already opened files
divided by IDHASHSIZE (256), which is O(n) if n is the number of opened
files.

The solution proposed changes the type of this variable to std::map and
makes complexity of all operations O(logn).

This change also improves readability of this part of the code.